修复运行时因为提前删除项目文件导致上传失败

+++ src/main/kotlin/github/zimo/autojsx/action/command/SaveAndRunProject.kt
@@ -40,7 +40,7 @@ class SaveAndRunProject :
 
             val zip = File(project?.basePath + "/build-output" + "/${name}.zip")
             zip.parentFile.mkdirs()
-            zip.delete()
+            if (zip.exists()) zip.delete()
 
             executor.submit {
                 zip(
@@ -49,8 +49,11 @@ class SaveAndRunProject :
                 )
                 VertxServer.Command.saveProject(zip.canonicalPath)
                 VertxServer.Command.runProject(zip.canonicalPath)
-                zip.delete()
                 logI("项目正在上传: " + projectJson.path)
+                logI("正在上传 src: " + src.path)
+                logI("项目正在上传 resources: " + resources.path)
+                logI("项目正在上传 lib: " + lib.path+"\r\n")
+//                if (zip.exists()) zip.delete()
             }
             return
         }
